feat: UX improvements - Select2 tenant search, building-grouped units, deposit termination modal, daily rent reminders, maintenance cascading dropdown, EGP currency, labeled buttons

+++ app/Services/RentScheduleService.php
@@ -71,7 +71,7 @@ class RentScheduleService
         $periods   = [];
         $current   = Carbon::parse($contract->start_date);
         $end       = Carbon::parse($contract->end_date);
-        $dueDay    = $contract->due_day ?? $current->day;
+        $dueDay    = (int) ($contract->due_day ?? $current->day);
         $interval  = $this->cycleInterval($contract->payment_cycle);
 
         while ($current->lte($end)) {
